[feat] 타워 건설 ui
타워 건설과 업그레이드 기능 구현
